Move quote creation form to new route and filter listing
